requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# CLI
typer==0.9.0
rich==13.7.0

# Database
sqlalchemy==2.0.36
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]
pydantic-settings

# Background Tasks
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Email Services
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
sendgrid==6.11.0

# AI Integration
openai==1.10.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
pytz==2023.4
pendulum==3.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
cryptography==42.0.0
PyJWT==2.10.1
requests==2.32.5
requests-oauthlib==2.0.0
